# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: intake
    description: Intake form operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /intake:
    post:
      operationId: createIntakeSubmission
      tags: [intake]
      summary: Submit an intake request
      description: Stores a new intake form submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntakeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeSubmission"
  /intake/submissions:
    get:
      operationId: listIntakeSubmissions
      tags: [intake]
      summary: List intake submissions
      description: Returns all intake submissions, newest first. Requires admin authorization.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IntakeSubmission"
        "401":
          description: Unauthorized
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    IntakeInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 320
        mobile:
          type: ["string", "null"]
          maxLength: 40
        company:
          type: ["string", "null"]
          maxLength: 200
        role:
          type: ["string", "null"]
          maxLength: 200
        companySize:
          type: string
          enum:
            - "Under 50"
            - "50-250"
            - "250-1,000"
            - "1,000-5,000"
            - "5,000+"
        projectDetails:
          type: string
          minLength: 1
          maxLength: 5000
      required:
        - name
        - email
        - companySize
        - projectDetails
    IntakeSubmission:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        mobile:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        companySize:
          type: string
          enum:
            - "Under 50"
            - "50-250"
            - "250-1,000"
            - "1,000-5,000"
            - "5,000+"
        projectDetails:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - companySize
        - projectDetails
        - createdAt